+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.57] - 2020-11-14
+### Added
++ SPM and LLB Builders.
++ Toolchain now contains `swift-build` binary.
+### Changed
+* Fixed error "unknown argument: '-modulewrap'" when building sources using Swift Package Manager.
+
 ## [1.0.56] - 2020-11-10
 ### Changed
 * Build scripts converted from Ruby to JS